# Send hello and login asynchronously to speed up connecting