Current we send HelloRequest, wait for a response, than send ConnectRequest and wait for a response.

We can send the `HelloRequest` and `ConnectRequest`, and than wait for both responses instead as it reduces one round trip wait. 

This should speed up connecting with encryption enabled a little bit since it reported to be slow in https://github.com/home-assistant/core/issues/103517#issuecomment-1798017912

```
INFO Starting log output from largegarage.local using esphome API
INFO Successfully connected to largegarage in 0.061s
INFO Successful handshake with largegarage in 0.022s
[18:45:21][I][app:102]: ESPHome version 2023.11.0b2 compiled on Nov  8 2023, 23:13:11
```
```
INFO Starting log output from icemaker.local using esphome API
INFO Successfully connected to icemaker in 0.020s
INFO Successful handshake with icemaker in 0.045s
[18:46:15][I][app:102]: ESPHome version 2023.12.0-dev compiled on Nov  9 2023, 16:50:08
```