@bdraco 2023-11-10
Manually test with wrong encryption key as well to make sure it doesn't cause a problem